Reviewer's comment 1 Fig 1:
I struggled with ‘Tleaf sensitivity to shortwave radiation’ but I think it is clear after, when reading the text. Consider using the same words in the paragraph where you detail it (line 268?) and adding the reference to the figure in the text. Or maybe change the term (‘thermal capacitance’ ?) if this is the part you are referring to in the text.