@teixeirak updated this sentence
Not sure if R2 read it as just respiration instead of Q10 but I added this sentence, let me know if it makes sense to the comment:

R2's comment: we currently lack mechanistic understanding of how and why the temperature sensitivity of respiration varies across the forest vertical profile'. Other than that it varies according to the prevailing temperature difference? I think the temperature difference between upper and lower canopy positions, the theme of this review, should be mentioned in regard to this